Rewrote force pause, made MinZ customizable
Force pause now being monitored by game engine, instead of cheat engine monitoring for hotkeys, so it should be much more responsive.  MinZ is customizable now; setting depth multiplier to 1 (default) is the same as the old function, 2 is the same as previous "aggressive mode."  1.5 is a much better compromise.